Relocate raw_query
