+++ server/repository/src/db_diesel/mod.rs
@@ -240,6 +240,8 @@ use diesel::{
     prelude::*,
     r2d2::{ConnectionManager, Pool, PooledConnection},
     result::{DatabaseErrorKind as DieselDatabaseErrorKind, Error as DieselError},
+    sql_query,
+    sql_types::Text,
 };
 
 #[cfg(not(feature = "postgres"))]
@@ -311,4 +313,16 @@ fn get_connection(
         msg: "Failed to open Connection".to_string(),
         extra: format!("{:?}", error),
     })
+}
+
+#[derive(QueryableByName, Debug, PartialEq)]
+pub struct JsonRawRow {
+    #[diesel(sql_type = Text)]
+    pub json_row: String,
+}
+
+pub fn raw_query(connection: &StorageConnection, query: String) -> Vec<JsonRawRow> {
+    sql_query(&query)
+        .get_results::<JsonRawRow>(connection.lock().connection())
+        .unwrap()
 }
